[feature] template: advanced search 추가 - chip variant, rounded 추가 - button color 추가
